refactor(tools): extract contacts domain into src/tools/contacts.py
